Change ord() operations to be handled for Py2 and Py3
Since almost all operations are now on bytes() objects in Python 3,
running ord() will result in errors. This changes instances of ord() to
use the function to_i() instead which only considers the binary integer
representation. As such, comparisons are made into the binary format,
with explaining comments towards which ASCII character we are looking
for.