Wrap transport strings in six.b()
Since Python 2 and 3 have wildly different opinions on what strings are,
there are a lot of troubles to be concerned with in regards to when
strings are to be handled as binary blobs.

Wrapping all strings that are supposed to be considered binary inside of
six.b() makes sure that both 3 and 2 can handle them accordingly.